Inject OapSparkSqlParser by SparkSessionExtension (#1005)
* inject OapSparkSqlParser by SparkSessionExtension

* Remove SparkSqlParser

* Remove Extra configuration.

* adjust object to class
